+++ testData/org/elixir_lang/injection/markdown/blank_line_in_code_block.ex
@@ -1,0 +1,11 @@
+defmodule BlankLineInCodeBlock do
+  @moduledoc """
+  A paragraph before the code block.
+
+      iex> a = 1
+  
+      iex> b = 2
+  """
+
+  def example, do: :ok
+end